Upgrade setuptools in venv
Recent versions of the Monasca agent require a minimum of
setuptools 1.4. If an earlier version is installed the installation
of the Monasca agent fails, complaining that it can't upgrade
setuptools, and that version >=1.4 is required.

An alternative would be to support installing arbitrary pip
packages into the venv via a user configurable variable.