Workaround for custom Monasca Agent plugins
We need to install the Monasca Agent custom plugins in the same venv
as the Monasca Agent so that the required dependencies are available
when the Agent runs the plugins. However, the Monasca Agent searchs
a hard coded path for plugins outside of the venv. This change symlinks
the plugin folder from the venv to the hard coded path.

Eventually the Monasca Agent should be able to discover the plugins
from the venv.